bd_init/pyproject.toml
[tool.poetry]
name = "bd-init"
version = "0.1.0"
description = ""
authors = [" leonideliseev <leonid.2004eliseev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
